requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
faiss-cpu==1.7.4
openai==1.78.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
requests==2.31.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0 
pydantic[email]==2.4.2
numpy==1.26.1